Moved all the logic of Commands to StudentManager
